+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Removed `TickTock.remove_outliers`
 - Moved `is_windows` to static `OS` attribute, which also contains `is_mac` and `is_linux`
 - Added `HardwareInfo`, which contains a bunch of system hardware information
+- `Figure` now automatically creates necessary subdirectories
 
 ### Bug fixes